backend/requirements.txt
betterpath==0.2.2
certifi==2017.11.5
chardet==3.0.4
click==6.7
coreapi==2.3.3
coreschema==0.0.4
Django==1.11.7
django-cors-headers==2.1.0
django-crispy-forms==1.7.0
django-filter==1.1.0
django-formtools==2.1
django-guardian==1.4.9
django-reversion==2.0.10
djangorestframework==3.7.3
future==0.16.0
httplib2==0.10.3
idna==2.6
itypes==1.1.0
Jinja2==2.10
l==0.11.0
Markdown==2.6.9
MarkupSafe==1.0
olefile==0.44
Pillow==4.3.0
PyMySQL==0.7.11
pytz==2017.3
requests==2.18.4
six==1.11.0
uritemplate==3.0.0
urllib3==1.22
vcversioner==2.16.0.0
XlsxWriter==1.0.2
xlwt==1.3.0
zope.interface==4.4.3
